Fixing deletion on 1st asset
